Split interval and time modules
